move TransitionLink for proper folder
