+++ components/post/PostRedirectLink.tsx
@@ -2,7 +2,7 @@
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { IconDefinition } from "@fortawesome/fontawesome-svg-core";
-import TransitionLink from "@/components/TransitionLink";
+import TransitionLink from "@/components/general/TransitionLink";
 import { Post } from "@/utils/types";
 
 interface PostRedirectLinkProps {
